extract dedicated form theme for minimum amount and show it in one row
