+++ src/PrestaShopBundle/Resources/config/services/core/form/choice_provider.yml
@@ -323,6 +323,10 @@ services:
     autowire: true
     public: false
 
+  PrestaShop\PrestaShop\Core\Form\ChoiceProvider\ShippingInclusionChoiceProvider:
+    autowire: true
+    public: false
+
   PrestaShop\PrestaShop\Core\Form\ChoiceProvider\CurrencyByIdChoiceProvider:
     autowire: true
     public: false
